Move OpSet to separate file
